modules/python/requirements.txt
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
pathlib2==2.3.7.post1
pillow==11.1.0
pytest==8.3.4
requests==2.32.3
svg.py==1.5.0
